requirements.txt
numpy
colorthief
scikit-learn
Pillow
colormath
wand